Add files via upload
Excel File that had a few adjustments made in Excel, used in the pre-processing code.  Note that I doubled up the code in the sense that each individual fight has been placed as Fighter A and as Fighter B to allow for the creation of a upcoming prediction file.  So this means that the file is twice as large the initially scrapped information!